little bit of reorganizing things
