+++ apps/web/src/components/settings/billing.tsx
@@ -1,19 +1,26 @@
 import { Suspense, useMemo, useState } from "react";
-import { 
-  usePlan, 
+import {
+  type PlanWithTeamMetadata,
+  usePlan,
   useWorkspaceWalletBalance,
   WELL_KNOWN_PLANS,
-  type PlanWithTeamMetadata 
 } from "@deco/sdk";
 import { Button } from "@deco/ui/components/button.tsx";
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@deco/ui/components/card.tsx";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@deco/ui/components/card.tsx";
 import { Icon } from "@deco/ui/components/icon.tsx";
 import { Skeleton } from "@deco/ui/components/skeleton.tsx";
 import { Badge } from "@deco/ui/components/badge.tsx";
 import { Separator } from "@deco/ui/components/separator.tsx";
 import { Alert, AlertDescription } from "@deco/ui/components/alert.tsx";
 import { Spinner } from "@deco/ui/components/spinner.tsx";
-import { 
+import { Progress } from "@deco/ui/components/progress.tsx";
+import {
   Dialog,
   DialogContent,
   DialogHeader,
@@ -36,7 +43,7 @@ const mockTransactions = [
     status: "completed" as const,
   },
   {
-    id: "tx_2", 
+    id: "tx_2",
     date: "2025-01-14",
     description: "Wallet Top-up",
     type: "topup" as const,
@@ -45,7 +52,7 @@ const mockTransactions = [
   },
   {
     id: "tx_3",
-    date: "2025-01-10", 
+    date: "2025-01-10",
     description: "Monthly Subscription - Growth Plan",
     type: "subscription" as const,
     amount: -2500,
@@ -100,7 +107,7 @@ const availablePlans: PlanInfo[] = [
     support: "Community support",
   },
   {
-    id: "starter", 
+    id: "starter",
     name: "Starter",
     price: 500,
     period: "month",
@@ -118,7 +125,7 @@ const availablePlans: PlanInfo[] = [
   },
   {
     id: "growth",
-    name: "Growth", 
+    name: "Growth",
     price: 2500,
     period: "month",
     credits: 2250,
@@ -140,7 +147,7 @@ const availablePlans: PlanInfo[] = [
     id: "scale",
     name: "Scale",
     price: 5000,
-    period: "month", 
+    period: "month",
     credits: 4750,
     seats: 100,
     markup: 5,
@@ -160,15 +167,17 @@ function WalletBalanceCard() {
   const { balance, refetch, isRefetching } = useWorkspaceWalletBalance();
 
   return (
-    <Card className="p-6 rounded-xl border">
-      <CardContent className="p-0">
+    <Card className="p-6 rounded-xl border h-full">
+      <CardContent className="p-0 h-full flex flex-col">
         <div className="flex items-center justify-between mb-4">
           <div className="flex items-center gap-2">
             <Icon name="account_balance_wallet" size={20} />
-            <span className="text-sm font-medium text-muted-foreground">Remaining Balance</span>
+            <span className="text-sm font-medium text-muted-foreground">
+              Remaining Balance
+            </span>
           </div>
-          <Button 
-            variant="ghost" 
+          <Button
+            variant="ghost"
             size="sm"
             onClick={refetch}
             disabled={isRefetching}
@@ -177,12 +186,12 @@ function WalletBalanceCard() {
             <Icon name="refresh" size={16} />
           </Button>
         </div>
-        
-        <div className="text-4xl font-semibold text-foreground mb-4">
+
+        <div className="text-5xl font-semibold text-foreground mb-4">
           {balance}
         </div>
-        
-        <div className="flex gap-2">
+
+        <div className="flex gap-2 mt-auto">
           <div className="w-fit">
             <DepositDialog />
           </div>
@@ -195,33 +204,12 @@ function WalletBalanceCard() {
   );
 }
 
-function BillingInfoCard() {
-  return (
-    <Card className="p-6 rounded-xl border">
-      <CardContent className="p-0">
-        <div className="flex items-center gap-2 mb-4">
-          <Icon name="receipt" size={20} />
-          <span className="text-sm font-medium text-muted-foreground">Billing Information</span>
-        </div>
-        
-        <div className="grid grid-cols-2 gap-4 text-sm">
-          <div>
-            <div className="text-muted-foreground">Billing Cycle</div>
-            <div className="font-medium">Monthly</div>
-          </div>
-          <div>
-            <div className="text-muted-foreground">Next Payment</div>
-            <div className="font-medium">August 1, 2025</div>
-          </div>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
 function PlanInfoCard() {
   const [showPlansModal, setShowPlansModal] = useState(false);
-  const currentPlan = availablePlans.find(p => p.current) || availablePlans[2];
+  const plan = usePlan();
+  // Calculate used seats
+  const usedSeats = plan.user_seats - plan.remainingSeats;
+  const seatLimit = plan.user_seats;
 
   return (
     <>
@@ -229,35 +217,47 @@ function PlanInfoCard() {
         <CardContent className="p-0 h-full flex flex-col">
           <div className="flex items-center justify-between mb-4">
             <div className="flex items-center gap-2">
-              <Icon name="workspace_premium" size={20} className="text-muted-foreground" />
-              <span className="text-sm font-medium text-muted-foreground">Current Plan</span>
+              <Icon
+                name="workspace_premium"
+                size={20}
+                className="text-muted-foreground"
+              />
+              <span className="text-sm font-medium text-muted-foreground">
+                Current Plan
+              </span>
             </div>
             <Badge variant="secondary" className="text-xs">
-              {currentPlan.name}
+              {plan.title}
             </Badge>
           </div>
-          
+
           <div className="space-y-3 text-sm flex-1">
-            <div className="flex justify-between">
-              <span className="text-muted-foreground">Seats used</span>
-              <span className="font-medium">21/50</span>
+            <div className="space-y-2">
+              <div className="flex justify-between">
+                <span className="text-muted-foreground">Seats used</span>
+                <span className="font-medium">{usedSeats}/{seatLimit}</span>
+              </div>
+              <Progress
+                value={(usedSeats / seatLimit) * 100}
+                className="h-2"
+              />
             </div>
             <div className="flex justify-between">
               <span className="text-muted-foreground">Monthly cost</span>
-              <span className="font-medium">${currentPlan.price/100}</span>
-            </div>
-            <div className="flex justify-between">
-              <span className="text-muted-foreground">Tech support</span>
-              <span className="font-medium">1h/week</span>
+              <span className="font-medium">
+                ${plan.monthly_credit_in_dollars}
+              </span>
             </div>
             <div className="flex justify-between">
               <span className="text-muted-foreground">AI usage rate</span>
-              <span className="font-medium text-primary">{currentPlan.markup}% markup</span>
+              <span className="font-medium text-primary">
+                {plan.markup}% markup
+              </span>
             </div>
           </div>
-          
-          <Button 
-            variant="outline" 
+
+          <Button
+            variant="outline"
             className="w-full mt-4"
             onClick={() => setShowPlansModal(true)}
           >
@@ -267,15 +267,17 @@ function PlanInfoCard() {
         </CardContent>
       </Card>
 
-      <PlansModal 
-        isOpen={showPlansModal} 
-        onClose={() => setShowPlansModal(false)} 
+      <PlansModal
+        isOpen={showPlansModal}
+        onClose={() => setShowPlansModal(false)}
       />
     </>
   );
 }
 
-function PlansModal({ isOpen, onClose }: { isOpen: boolean; onClose: () => void }) {
+function PlansModal(
+  { isOpen, onClose }: { isOpen: boolean; onClose: () => void },
+) {
   const handlePlanSelect = (planId: string) => {
     console.log(`Selected plan: ${planId}`);
     onClose();
@@ -287,25 +289,27 @@ function PlansModal({ isOpen, onClose }: { isOpen: boolean; onClose: () => void 
         <DialogHeader>
           <DialogTitle>Choose Your Plan</DialogTitle>
         </DialogHeader>
-        
+
         <div className="p-6">
           <div className="grid grid-cols-1 md:grid-cols-2 xl:grid-cols-4 gap-6">
             {availablePlans.map((plan) => (
-              <PlanCard 
-                key={plan.id} 
-                plan={plan} 
+              <PlanCard
+                key={plan.id}
+                plan={plan}
                 onSelect={handlePlanSelect}
               />
             ))}
           </div>
         </div>
-        
+
         {/* Custom Plan Card */}
         <Card className="p-6 border-dashed mx-6 mb-6">
           <CardContent className="p-0 text-center space-y-4">
             <div className="space-y-2">
               <h4 className="text-lg font-semibold">Custom Plan</h4>
-              <p className="text-sm text-muted-foreground">For organizations with specific needs</p>
+              <p className="text-sm text-muted-foreground">
+                For organizations with specific needs
+              </p>
             </div>
             <div className="text-sm space-y-1">
               <div className="flex items-center justify-center gap-2">
@@ -336,22 +340,28 @@ function PlansModal({ isOpen, onClose }: { isOpen: boolean; onClose: () => void 
   );
 }
 
-function PlanCard({ plan, onSelect }: { plan: PlanInfo; onSelect: (planId: string) => void }) {
+function PlanCard(
+  { plan, onSelect }: { plan: PlanInfo; onSelect: (planId: string) => void },
+) {
   const isPopular = plan.popular;
   const isCurrent = plan.current;
-  
+
   return (
-    <Card className={`relative overflow-hidden transition-all hover:shadow-lg min-h-[500px] flex flex-col ${
-      isCurrent ? "ring-2 ring-primary shadow-md" : ""
-    } ${isPopular ? "border-primary shadow-md scale-105" : ""}`}>
-      
+    <Card
+      className={`relative overflow-hidden transition-all hover:shadow-lg min-h-[500px] flex flex-col ${
+        isCurrent ? "ring-2 ring-primary shadow-md" : ""
+      } ${isPopular ? "border-primary shadow-md scale-105" : ""}`}
+    >
       {/* Header with gradient background */}
-      <div className={`relative px-6 py-6 text-center ${
-        isPopular ? "bg-gradient-to-br from-primary/10 to-primary/5" : 
-        isCurrent ? "bg-gradient-to-br from-muted/50 to-muted/20" : 
-        "bg-gradient-to-br from-muted/30 to-muted/10"
-      }`}>
-        
+      <div
+        className={`relative px-6 py-6 text-center ${
+          isPopular
+            ? "bg-gradient-to-br from-primary/10 to-primary/5"
+            : isCurrent
+            ? "bg-gradient-to-br from-muted/50 to-muted/20"
+            : "bg-gradient-to-br from-muted/30 to-muted/10"
+        }`}
+      >
         {/* Badges */}
         {isPopular && (
           <div className="absolute top-2 left-2">
@@ -367,40 +377,52 @@ function PlanCard({ plan, onSelect }: { plan: PlanInfo; onSelect: (planId: strin
             </Badge>
           </div>
         )}
-        
+
         {/* Plan icon */}
-        <div className={`mx-auto w-12 h-12 rounded-full flex items-center justify-center mb-4 ${
-          isPopular ? "bg-primary text-primary-foreground" : 
-          isCurrent ? "bg-muted text-muted-foreground" : 
-          "bg-muted/60 text-muted-foreground"
-        }`}>
-          <Icon name={plan.id === 'free' ? 'star_border' : 
-                       plan.id === 'starter' ? 'rocket_launch' : 
-                       plan.id === 'growth' ? 'trending_up' : 'enterprise'} 
-                size={20} />
+        <div
+          className={`mx-auto w-12 h-12 rounded-full flex items-center justify-center mb-4 ${
+            isPopular
+              ? "bg-primary text-primary-foreground"
+              : isCurrent
+              ? "bg-muted text-muted-foreground"
+              : "bg-muted/60 text-muted-foreground"
+          }`}
+        >
+          <Icon
+            name={plan.id === "free"
+              ? "star_border"
+              : plan.id === "starter"
+              ? "rocket_launch"
+              : plan.id === "growth"
+              ? "trending_up"
+              : "enterprise"}
+            size={20}
+          />
         </div>
-        
+
         {/* Plan name */}
         <h3 className="text-xl font-bold mb-2">{plan.name}</h3>
-        
+
         {/* Pricing */}
         <div className="flex items-baseline justify-center gap-1 mb-2">
-          <span className="text-4xl font-bold">${plan.price/100}</span>
+          <span className="text-4xl font-bold">${plan.price / 100}</span>
           <span className="text-muted-foreground text-sm">/{plan.period}</span>
         </div>
-        
+
         {/* Tagline */}
         <p className="text-sm text-muted-foreground">
-          {plan.id === 'free' ? 'Get started for free' :
-           plan.id === 'starter' ? 'Perfect for small teams' :
-           plan.id === 'growth' ? 'For growing businesses' :
-           'For large organizations'}
+          {plan.id === "free"
+            ? "Get started for free"
+            : plan.id === "starter"
+            ? "Perfect for small teams"
+            : plan.id === "growth"
+            ? "For growing businesses"
+            : "For large organizations"}
         </p>
       </div>
-      
+
       {/* Content */}
       <div className="flex-1 p-6 space-y-6">
-        
         {/* Key metrics */}
         <div className="grid grid-cols-2 gap-4">
           <div className="text-center p-3 bg-muted/20 rounded-lg">
@@ -408,28 +430,39 @@ function PlanCard({ plan, onSelect }: { plan: PlanInfo; onSelect: (planId: strin
             <div className="text-xs text-muted-foreground">Seats</div>
           </div>
           <div className="text-center p-3 bg-muted/20 rounded-lg">
-            <div className="text-2xl font-bold text-primary">${plan.credits/100}</div>
+            <div className="text-2xl font-bold text-primary">
+              ${plan.credits / 100}
+            </div>
             <div className="text-xs text-muted-foreground">AI Credits</div>
           </div>
         </div>
-        
+
         {/* Usage rate highlight */}
         <div className="text-center p-3 bg-gradient-to-r from-green-50 to-blue-50 rounded-lg border">
-          <div className="text-lg font-semibold text-green-700">{plan.markup}% markup</div>
+          <div className="text-lg font-semibold text-green-700">
+            {plan.markup}% markup
+          </div>
           <div className="text-xs text-muted-foreground">AI Usage Rate</div>
         </div>
 
         {/* Features */}
         <div className="space-y-3">
-          {plan.features.slice(0, 3).map((feature, index) => (
-            <div key={index} className="flex items-start gap-3">
+          {plan.features.slice(0, 3).map((feature) => (
+            <div
+              key={`${plan.id}-${feature.name}`}
+              className="flex items-start gap-3"
+            >
               <div className="w-5 h-5 rounded-full bg-green-100 flex items-center justify-center mt-0.5">
                 <Icon name="check" size={12} />
               </div>
               <div className="flex-1">
-                <div className="text-sm font-medium text-foreground">{feature.name}</div>
+                <div className="text-sm font-medium text-foreground">
+                  {feature.name}
+                </div>
                 {feature.limit && (
-                  <div className="text-xs text-muted-foreground">{feature.limit}</div>
+                  <div className="text-xs text-muted-foreground">
+                    {feature.limit}
+                  </div>
                 )}
               </div>
             </div>
@@ -439,15 +472,21 @@ function PlanCard({ plan, onSelect }: { plan: PlanInfo; onSelect: (planId: strin
 
       {/* Footer */}
       <div className="p-6 pt-0">
-        <Button 
+        <Button
           variant={isCurrent ? "secondary" : isPopular ? "default" : "outline"}
-          className={`w-full ${isPopular ? "bg-primary hover:bg-primary/90" : ""}`}
+          className={`w-full ${
+            isPopular ? "bg-primary hover:bg-primary/90" : ""
+          }`}
           disabled={isCurrent}
           onClick={() => onSelect(plan.id)}
         >
-          {isCurrent ? "Current Plan" : plan.price === 0 ? "Downgrade" : "Upgrade"}
+          {isCurrent
+            ? "Current Plan"
+            : plan.price === 0
+            ? "Downgrade"
+            : "Upgrade"}
         </Button>
-        
+
         {/* Support info */}
         <div className="text-center mt-2">
           <span className="text-xs text-muted-foreground">{plan.support}</span>
@@ -464,14 +503,14 @@ function TransactionsTable() {
   const getStatusBadge = (status: Transaction["status"]) => {
     const variants = {
       completed: "default",
-      pending: "secondary", 
+      pending: "secondary",
       failed: "destructive",
     } as const;
-    
+
     const labels = {
       completed: "Completed",
       pending: "Pending",
-      failed: "Failed", 
+      failed: "Failed",
     };
 
     return (
@@ -486,17 +525,17 @@ function TransactionsTable() {
       subscription: "workspace_premium",
       topup: "add_circle",
     };
-    
+
     const colors = {
       subscription: "text-blue-600",
-      topup: "text-green-600", 
+      topup: "text-green-600",
     };
 
     return (
-      <Icon 
-        name={icons[type]} 
-        size={16} 
-        className={colors[type]} 
+      <Icon
+        name={icons[type]}
+        size={16}
+        className={colors[type]}
       />
     );
   };
@@ -514,19 +553,21 @@ function TransactionsTable() {
       sortable: true,
     },
     {
-      id: "description", 
+      id: "description",
       header: "Description",
       render: (transaction) => (
         <div className="flex flex-col">
           <span className="font-medium">{transaction.description}</span>
-          <span className="text-xs text-muted-foreground capitalize">{transaction.type}</span>
+          <span className="text-xs text-muted-foreground capitalize">
+            {transaction.type}
+          </span>
         </div>
       ),
       sortable: true,
     },
     {
       id: "status",
-      header: "Status", 
+      header: "Status",
       render: (transaction) => getStatusBadge(transaction.status),
       sortable: true,
     },
@@ -534,34 +575,23 @@ function TransactionsTable() {
       id: "amount",
       header: "Amount",
       render: (transaction) => (
-        <span className={`font-medium ${
-          transaction.amount > 0 ? "text-green-600" : "text-foreground"
-        }`}>
-          {transaction.amount > 0 ? "+" : ""}${Math.abs(transaction.amount/100).toFixed(2)}
+        <span
+          className={`font-medium ${
+            transaction.amount > 0 ? "text-green-600" : "text-foreground"
+          }`}
+        >
+          {transaction.amount > 0 ? "+" : ""}${Math.abs(
+            transaction.amount / 100,
+          ).toFixed(2)}
         </span>
       ),
       sortable: true,
     },
   ];
 
-  const getSortValue = (transaction: Transaction, key: string): string | number => {
-    switch (key) {
-      case "date":
-        return new Date(transaction.date).getTime();
-      case "description":
-        return transaction.description.toLowerCase();
-      case "status":
-        return transaction.status;
-      case "amount":
-        return transaction.amount;
-      default:
-        return "";
-    }
-  };
-
   const handleSort = (key: string) => {
     if (sortKey === key) {
-      setSortDirection(prev => prev === "asc" ? "desc" : "asc");
+      setSortDirection((prev) => prev === "asc" ? "desc" : "asc");
     } else {
       setSortKey(key);
       setSortDirection("asc");
@@ -569,6 +599,24 @@ function TransactionsTable() {
   };
 
   const sortedTransactions = useMemo(() => {
+    const getSortValue = (
+      transaction: Transaction,
+      key: string,
+    ): string | number => {
+      switch (key) {
+        case "date":
+          return new Date(transaction.date).getTime();
+        case "description":
+          return transaction.description.toLowerCase();
+        case "status":
+          return transaction.status;
+        case "amount":
+          return transaction.amount;
+        default:
+          return "";
+      }
+    };
+
     return [...mockTransactions].sort((a, b) => {
       const aVal = getSortValue(a, sortKey);
       const bVal = getSortValue(b, sortKey);
@@ -579,7 +627,7 @@ function TransactionsTable() {
 
       const aStr = String(aVal);
       const bStr = String(bVal);
-      
+
       if (aStr < bStr) return sortDirection === "asc" ? -1 : 1;
       if (aStr > bStr) return sortDirection === "asc" ? 1 : -1;
       return 0;
@@ -590,7 +638,9 @@ function TransactionsTable() {
     <div className="space-y-4">
       <div className="flex items-center justify-between">
         <h3 className="text-lg font-semibold">Billing History</h3>
-        <p className="text-sm text-muted-foreground">Your subscription payments and wallet top-ups</p>
+        <p className="text-sm text-muted-foreground">
+          Your subscription payments and wallet top-ups
+        </p>
       </div>
       <Table
         columns={columns}
@@ -608,7 +658,8 @@ function BillingErrorFallback() {
     <Alert variant="destructive" className="my-8">
       <Icon name="error" size={16} />
       <AlertDescription>
-        Something went wrong while loading the billing data. Please try again later.
+        Something went wrong while loading the billing data. Please try again
+        later.
       </AlertDescription>
     </Alert>
   );
@@ -624,9 +675,6 @@ export default function BillingSettings() {
             <div className="space-y-6">
               <Suspense fallback={<Skeleton className="h-[200px]" />}>
                 <WalletBalanceCard />
-              </Suspense>
-              <Suspense fallback={<Skeleton className="h-[180px]" />}>
-                <BillingInfoCard />
               </Suspense>
             </div>
             <div>
@@ -644,4 +692,4 @@ export default function BillingSettings() {
       </ErrorBoundary>
     </div>
   );
-} 
+}
