+++ packages/sdk/src/hooks/wallet.ts
@@ -39,6 +39,9 @@ export function useUsagePerAgent({
   return usage;
 }
 
+export type AgentUsage = Awaited<ReturnType<typeof useUsagePerAgent>>;
+export type AgentUsageItem = AgentUsage["items"][number];
+
 export function useUsagePerThread({
   range,
 }: {
@@ -52,6 +55,9 @@ export function useUsagePerThread({
 
   return usage;
 }
+
+export type ThreadUsage = Awaited<ReturnType<typeof useUsagePerThread>>;
+export type ThreadUsageItem = ThreadUsage["items"][number];
 
 export function usePlan() {
   const { workspace } = useSDK();
